+++ src/DeveloperBadges.hpp
@@ -1,16 +1,7 @@
 #include <ccTypes.h>
-
-enum class BadgeType {
-    None,
-    ModDeveloper,
-    VerifiedDeveloper,
-    IndexStaff,
-    LeadDeveloper
-};
-
 struct DeveloperBadge {
     int id;
-    BadgeType badge;
+    int type;
     std::string name;
 };
 
@@ -18,7 +9,6 @@ class DeveloperBadges {
 public:
     static inline std::vector<DeveloperBadge> developerBadges;
 
-    static DeveloperBadge& badgeForUser(int);
-    static cocos2d::ccColor3B getCommentColor(BadgeType);
-    static void showBadgeInfo(const std::string&, BadgeType);
+    static DeveloperBadge* badgeForUser(int);
+    static void showBadgeInfo(const std::string&, int);
 };