+++ src/hooks/ProfilePage.cpp
@@ -6,6 +6,7 @@ using namespace geode::prelude;
 
 class $modify(DBProfilePage, ProfilePage) {
     struct Fields {
+        std::string m_badgeName;
         bool m_hasBadge;
     };
 
@@ -16,18 +17,20 @@ class $modify(DBProfilePage, ProfilePage) {
         if (f->m_hasBadge) return;
 
         auto badge = DeveloperBadges::badgeForUser(score->m_accountID);
-        if (badge.id == 0) return;
+        if (!badge) return;
 
         auto usernameMenu = m_mainLayer->getChildByID("username-menu");
         if (!usernameMenu) return;
 
+        f->m_badgeName = badge->name;
+
+        auto badgeType = badge->type;
         auto badgeButton = CCMenuItemSpriteExtra::create(
-            CCSprite::createWithSpriteFrameName(fmt::format("badge{:02}.png"_spr, (int)badge.badge).c_str()),
+            CCSprite::createWithSpriteFrameName(fmt::format("badge{:02}.png"_spr, badgeType).c_str()),
             this, menu_selector(DBProfilePage::onBadge)
         );
         badgeButton->setID("developer-badge"_spr);
-        badgeButton->setTag((int)badge.badge);
-        badgeButton->setUserObject("badge-name"_spr, CCString::create(badge.name));
+        badgeButton->setTag(badgeType);
         usernameMenu->addChild(badgeButton);
         usernameMenu->updateLayout();
 
@@ -35,7 +38,6 @@ class $modify(DBProfilePage, ProfilePage) {
     }
 
     void onBadge(CCObject* sender) {
-        auto badgeName = static_cast<CCString*>(static_cast<CCNode*>(sender)->getUserObject("badge-name"_spr));
-        DeveloperBadges::showBadgeInfo(badgeName ? badgeName->m_sString : m_score->m_userName, (BadgeType)sender->getTag());
+        DeveloperBadges::showBadgeInfo(m_fields->m_badgeName, sender->getTag());
     }
 };